requirements.txt
requests>=2.31.0
firebase-admin>=6.2.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
Pillow>=10.0.0
Flask>=2.3.0
Werkzeug>=2.3.0
gunicorn>=21.2.0
